requirements.txt
langchain
langchain-groq
langchain-community
chromadb
sentence-transformers
pypdf
python-dotenv